+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - [feat:support tsf unit.](https://github.com/Tencent/spring-cloud-tencent/pull/1723)
 - [feat: support service registry and discovery with Polaris and Nacos](https://github.com/Tencent/spring-cloud-tencent/pull/1728)
 - [test:add junit tests to sct-common.](https://github.com/Tencent/spring-cloud-tencent/pull/1730)
+- [test:add junit tests to sct-gw-plugin.](https://github.com/Tencent/spring-cloud-tencent/pull/1732)